Fix: 로그아웃 부분 버그 수정
RT가 유효한지 체크하는 과정에서
모종의 사유로 FE가 RT를 제공하지 않았을 경우에
로그아웃이 되지 않는 문제를 해결

로깅을 하는 것으로 대체하였음
해당 에러가 발생할 경우 원인을 찾아 수정하여야 함

기타 수정

+++ controllers/auth/login.auth.controller.js
@@ -58,9 +58,9 @@ const logout = async (req, res, next) => {
     // 1. RT가 DB에 존재하는지 확인하고, 존재한다면 DB에서 제거
     // 1-1. DB에 존재하지 않으면 에러 로깅
     const refreshToken = req.cookies.SPECTOGETHER_RT;
-    tokenService.checkIfRefreshTokenExists(refreshToken);
+    const isTokenExist = tokenService.checkIfRefreshTokenExists(refreshToken);
     // 1-2. DB에서 제거
-    await loginService.deleteRefreshToken(refreshToken);
+    if (isTokenExist) await loginService.deleteRefreshToken(refreshToken);
 
     // 2. 응답
     return res
